Add test showing diff proof of concept
Computes the diff output using an existing package and outputs it within
the test. Run the test with,

```sh
go test -v ./pkg/models -run=TestDashboardDiff
```

Outputting the diff in raw bytes probably doesn't help us at all, but
it's a start. It's also likely that the package I imported to compute
the diff wouldn't be acceptable in Grafana core due to it's usage of the
Godeps package management tool.